+++ src/orquestra/sdk/kubernetes/__init__.py
@@ -1,0 +1,2 @@
+# This module is vendored from https://github.com/kubernetes-client/python
+# See each source file for copyright.